fix: update auth tests for /dashboard redirect and polish empty state UI
