+++ src/components/BackButton.tsx
@@ -6,7 +6,7 @@ export default function BackButton({ label = "Go Back" }: { label?: string }) {
 
     return (
         <button
-            onClick={() => router.push("/dashboard")}
+            onClick={() => router.push("/")}
             className="border border-white bg-black-600 text-white px-4 py-2 rounded hover:bg-gray-700 active:scale-95 transition-all"
         >
             {label}
